Move account-related views to own module
